Added correct role allowed endpoints

+++ core/src/main/java/sit/tu_varna/bg/core/operationservice/review/DeleteMyReviewService.java
@@ -3,25 +3,25 @@ package sit.tu_varna.bg.core.operationservice.review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.transaction.Transactional;
 import sit.tu_varna.bg.api.exception.ResourceNotFoundException;
-import sit.tu_varna.bg.api.operation.review.delete.DeleteReviewOperation;
-import sit.tu_varna.bg.api.operation.review.delete.DeleteReviewRequest;
-import sit.tu_varna.bg.api.operation.review.delete.DeleteReviewResponse;
+import sit.tu_varna.bg.api.operation.review.deletemy.DeleteMyReviewOperation;
+import sit.tu_varna.bg.api.operation.review.deletemy.DeleteMyReviewRequest;
+import sit.tu_varna.bg.api.operation.review.deletemy.DeleteMyReviewResponse;
 import sit.tu_varna.bg.entity.Review;
 
 import java.util.UUID;
 
 @ApplicationScoped
-public class DeleteReviewService implements DeleteReviewOperation {
+public class DeleteMyReviewService implements DeleteMyReviewOperation {
 
     @Transactional
     @Override
-    public DeleteReviewResponse process(DeleteReviewRequest request) {
+    public DeleteMyReviewResponse process(DeleteMyReviewRequest request) {
         UUID reviewId = request.getReviewId();
         Review review = (Review) Review.findByIdOptional(reviewId)
                 .orElseThrow(() -> new ResourceNotFoundException("Review with id " + reviewId + " not found"));
-        if (review.isPersistent()) {
+        if (review.isPersistent() && review.getUser().getId().equals(request.getUserId())) {
             review.delete();
         }
-        return DeleteReviewResponse.builder().deleted(true).build();
+        return DeleteMyReviewResponse.builder().deleted(true).build();
     }
 }
